fix impl playcount dec
