+++ src/main/scala/coop/rchain/repo/Cache.scala
@@ -9,38 +9,19 @@ import scalacache.serialization.binary._
 import scalacache.memoization._
 import scalacache.modes.try_._
 
-import scala.util
-import scala.util.{Either, Failure, Left, Right, Success, Try}
+import scala.util.{Either, Left, Right, Try}
 
-object ErrImplicits {
-  implicit class _Either_[T](t: Try[T]) {
-    def asErr: Either[Err, T] = {
-      t match {
-        case Success(s) => Right(s)
-        case Failure(e) =>
-          Left(Err(ErrorCode.cacheLayer, e.getMessage, None))
-      }
-    }
-  }
-  implicit class _Try_[E,T](e: Either[E,T]) {
-    def asTry: Try[T] = {
-      e match {
-        case Right(s) => Success(s)
-        case Left(f) =>
-          util.Failure(CachingException(f.toString))
-      }
-    }
-  }
-}
+
 object RSongCache {
-  import ErrImplicits._
+  import coop.rchain.utils.ErrImplicits._
 
   case class CachedAsset(
     name: String,
     binaryData: Array[Byte]
   )
-  val log = Logger("RSongCache")
-  val repo = new SongRepo(Globals.proxy)
+  val log = Logger[RSongCache.type]
+  val  binaryAsset: String => Either[Err, Array[Byte]] = name =>
+    SongRepo.getRSongAsset(name)(Globals.proxy)
 
   val (redisUrl, redisPort)=
     (Globals.appCfg.getString("redis.url"),
@@ -50,12 +31,12 @@ object RSongCache {
     RedisCache(redisUrl, redisPort)
 
   rsongCache.config
-val getMemoizedAsset: String => SongRepo => Either[Err, CachedAsset] =
-  name => repo => {
+val getMemoizedAsset: String => RholangProxy => Either[Err, CachedAsset] =
+  name => proxy => {
 
   def  __getMemoizedAsset(name: String): Try[CachedAsset] =
     memoize[Try, CachedAsset](None) {
-      loadCache(name)(repo) match {
+      SongRepo.getRSongAsset(name)(proxy).map(CachedAsset(name, _)) match {
         case Right(s) => s
         case Left(e) =>
           log.error(s"Exception in RSongCache layer. ${e}")
@@ -66,9 +47,5 @@ val getMemoizedAsset: String => SongRepo => Either[Err, CachedAsset] =
     log.info(s"in memoized, attempting to fetch $name")
     __getMemoizedAsset(name).asErr
   }
-
-  val loadCache: String =>
-    SongRepo =>  Either[Err,CachedAsset]= name => repo =>
-    repo.getBinaryData(name).map(CachedAsset(name, _))
 
 }
